move contents of server folder to root
